Completion: Stage 2 (Add & config)
Update README.md to add second stage

config() now writes a user name to config.txt. It'll read the name from
config.txt or print a message if it is empty.

add() will print a message if index.txt is empty or the tracked files if it
isn't. Also adds a file to the list if it exists or an error message if it
doesn't.