Completion: Stage 3 (log & commit)
Main.kt now just handles reading the command line and calling the right
function in the new Operations class. Functions receive a list of Strings
arguments passed in excluding the command itself.

New Operations class to handle all the grunt work. The class also ensures
the vcs directory and necessary files are created.

Update the tests for the latest stage.